Splitt ut FagsakDeltager fra FagsakService/Controller
Et forsøk på å gjøre fagsak-klassene litt mindre. FagsakDeltager er kun knyttet opp mot søk. Derfor er det passende å trekke det ut til en egen controller og service.

Det er noe overlapp med endepunktene som returnerer RestFagsakIdOgTilknyttetAktørId som også bruker "/søk"-pathen. Har valgt å ikke dra dette inn, og så får vi heller trekke det ut i en annen runde. Der kan det være mer aktuelt å endre på endepunkt-pathen. Man kan vel se på det som et oppslag fremfor et søk.
